flask_backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
paho-mqtt==1.6.1
boto3==1.28.85
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7
marshmallow==3.20.1
flasgger==0.9.7.1
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
faker==19.6.2